Remove animation from lazy loading placeholders
This is a speculative fix for an issue which causes the background image used by the placeholder psuedo-element to be re-downloaded repeatedly by Blink based browsers. It appears that many events can cause the browser to perform a style recalculation which triggers the image download. I have not been able to recreate this issue locally nor produce an isolated test case from scratch, however I have been able to remove all elements from the live site except for a single placeholder. I am fairly certain this is a browser bug rather than created by us due to its inconsistency and lack of traceability.